Added a pure TensorFlow2 test which doesn't use Horovod. Also created a standard hook for applying the module, environment and system combinations discovered by find_modules.
